Implement fixes based on PR comments.
