+++ .gitignore
@@ -30,7 +30,6 @@ parts/
 sdist/
 var/
 wheels/
-venv/
 *.egg-info/
 .installed.cfg
 *.egg
