+++ src/skmultiflow/transform/windowed_minmax_scaler.py
@@ -4,14 +4,16 @@ from skmultiflow.transform.base_transform import StreamTransform
 from skmultiflow.utils import FastBuffer, get_dimensions
 
 
-class StandardScaler(StreamTransform):
-    """ Standardize features by removing the mean and scaling to unit variance.
-    Mean and stdev are computed in given window frame.
+class WindowedMinmaxScaler(StreamTransform):
+    """ Transform features by scaling each feature to a given range.
+    This estimator scales and translates each feature individually such
+    that it is in the given range on the training set, e.g. between zero and one.
+    For the training set we consider a window of a given length.
 
     Parameters
     ----------
     window_size: int (Default: 200)
-        Defines the window size to compute mean and standard deviation.
+        Defines the window size to compute min and max values.
 
     Examples
     --------
@@ -28,9 +30,7 @@ class StandardScaler(StreamTransform):
         self.window = FastBuffer(max_size=self.window_size)
 
     def transform(self, X):
-        """ transform
-
-        Does the transformation process in the samples in X.
+        """ Does the transformation process in the samples in X.
 
         Parameters
         ----------
@@ -43,31 +43,31 @@ class StandardScaler(StreamTransform):
             row = np.copy([X[i][:]])
             for j in range(c):
                 value = X[i][j]
-                mean = self._get_mean(j)
-                stdev = self._get_std(j)
-                transformed = (value - mean) / stdev
+                min_val = self._get_min(j)
+                max_val = self._get_max(j)
+                if((max_val-min_val)==0):
+                    transformed=0
+                else:
+                    X_std = (value - min_val) / (max_val - min_val)
+                    transformed = X_std * (max_val - min_val) + min_val
                 X[i][j] = transformed
             self.window.add_element(row)
         return X
 
-    def _get_mean(self, column_index):
-        mean = 0.
+    def _get_min(self, column_index):
+        min_val = 0.
         if not self.window.is_empty():
-            mean = np.nanmedian(np.array(self.window.get_queue())[:, column_index])
-        return mean
+            min_val = np.nanmin(np.array(self.window.get_queue())[:, column_index])
+        return min_val
 
-    def _get_std(self, column_index):
-        std = 1.
+    def _get_max(self, column_index):
+        max_val = 1.
         if not self.window.is_empty():
-            std = np.nanstd(np.array(self.window.get_queue())[:, column_index])
-        if(std == 0.):
-            std = 1.
-        return std
+            max_val = np.nanmax(np.array(self.window.get_queue())[:, column_index])
+        return max_val
 
     def partial_fit_transform(self, X, y=None):
-        """ partial_fit_transform
-
-        Partially fits the model and then apply the transform to the data.
+        """ Partially fits the model and then apply the transform to the data.
 
         Parameters
         ----------
@@ -88,9 +88,7 @@ class StandardScaler(StreamTransform):
         return X
 
     def partial_fit(self, X, y=None):
-        """ partial_fit
-
-        Partial fits the model.
+        """ Partial fits the model.
 
         Parameters
         ----------
@@ -102,10 +100,9 @@ class StandardScaler(StreamTransform):
 
         Returns
         -------
-        StandardScaler
+        MinmaxScaler
             self
 
         """
-        X = np.asarray(X)
         self.window.add_element(X)
         return self
