+++ src/skmultiflow/transform/windowed_standard_scaler.py
@@ -4,16 +4,14 @@ from skmultiflow.transform.base_transform import StreamTransform
 from skmultiflow.utils import FastBuffer, get_dimensions
 
 
-class MinmaxScaler(StreamTransform):
-    """ Transform features by scaling each feature to a given range.
-    This estimator scales and translates each feature individually such
-    that it is in the given range on the training set, e.g. between zero and one.
-    For the training set we consider a window of a given length.
+class WindowedStandardScaler(StreamTransform):
+    """ Standardize features by removing the mean and scaling to unit variance.
+    Mean and stdev are computed in given window frame.
 
     Parameters
     ----------
     window_size: int (Default: 200)
-        Defines the window size to compute min and max values.
+        Defines the window size to compute mean and standard deviation.
 
     Examples
     --------
@@ -30,9 +28,7 @@ class MinmaxScaler(StreamTransform):
         self.window = FastBuffer(max_size=self.window_size)
 
     def transform(self, X):
-        """ transform
-
-        Does the transformation process in the samples in X.
+        """ Does the transformation process in the samples in X.
 
         Parameters
         ----------
@@ -45,33 +41,29 @@ class MinmaxScaler(StreamTransform):
             row = np.copy([X[i][:]])
             for j in range(c):
                 value = X[i][j]
-                min = self._get_min(j)
-                max = self._get_max(j)
-                if((max-min)==0):
-                    transformed=0
-                else:
-                    X_std = (value - min) / (max - min)
-                    transformed = X_std * (max - min) + min
+                mean = self._get_mean(j)
+                stdev = self._get_std(j)
+                transformed = (value - mean) / stdev
                 X[i][j] = transformed
             self.window.add_element(row)
         return X
 
-    def _get_min(self, column_index):
-        min = 0.
+    def _get_mean(self, column_index):
+        mean = 0.
         if not self.window.is_empty():
-            min = np.nanmin(np.array(self.window.get_queue())[:, column_index])
-        return min
+            mean = np.nanmean(np.array(self.window.get_queue())[:, column_index])
+        return mean
 
-    def _get_max(self, column_index):
-        max = 1.
+    def _get_std(self, column_index):
+        std = 1.
         if not self.window.is_empty():
-            max = np.nanmax(np.array(self.window.get_queue())[:, column_index])
-        return max
+            std = np.nanstd(np.array(self.window.get_queue())[:, column_index])
+        if(std == 0.):
+            std = 1.
+        return std
 
     def partial_fit_transform(self, X, y=None):
-        """ partial_fit_transform
-
-        Partially fits the model and then apply the transform to the data.
+        """ Partially fits the model and then apply the transform to the data.
 
         Parameters
         ----------
@@ -92,9 +84,7 @@ class MinmaxScaler(StreamTransform):
         return X
 
     def partial_fit(self, X, y=None):
-        """ partial_fit
-
-        Partial fits the model.
+        """ Partial fits the model.
 
         Parameters
         ----------
@@ -106,10 +96,9 @@ class MinmaxScaler(StreamTransform):
 
         Returns
         -------
-        MinmaxScaler
+        StandardScaler
             self
 
         """
-        X = np.asarray(X)
         self.window.add_element(X)
         return self
